Add skeleton of per core fault handling
WIP! The handlers are not yet registered.
Each core gets its own fault handlers. The fault handlers get the
core's context pointer from a memory location.
We could lookup the core id in the fault handler via the APIC registers,
but this will most likely involve some microcode. If a fault handler
is entered, there is a good chance the microcode is still the custom
update. This might prevent APIC functions from working.
A workaround is to copy the fault handlers to a core specific buffer,
reference a memory location relative to the handlers for the core
context and then set the context pointer after copying the whole
structure.

The fault handlers for now only write the fault context to the
core context and then endless loop. Core 0 can then read the
fault context and report it to AngryCAT.

Core 0 will not get fault handling, don't fault core 0.